тесты функционала оценок

+++ database/migrations/2022_10_15_113418_create_ratings_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('ratings', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('client_id');
-            $table->tinyInteger('rating');
+            $table->float('rating');
             $table->morphs('ratingable');
             $table->timestamps();
         });
